Move test models out of test class for re-usability
